bring evictor files to top-level project for builds
